requirements.txt
uvicorn
fastapi
torch
python-shogi
numpy
pandas
pydantic